Refactor: move arc generator to a more general spot
* Move the method from the `drawing` module to the `geometry` module where it makes more sense.

* Consolidate a forgotten constant (QUART_PI) in `geometry`.
